disable certificate verification
I was unable to connect to the API without this change. The API did not respond to port 80 so I switched to SSL by prepending https:// to the host. This got me a self signed certificate error so in the end I decided to switch of the SSL certificate verification altogether.

This "fix" is not very pretty but since ExLibris decided to use a self-signed certificate I think this is the only way to use the API at the moment